Added flag owner time

+++ Hold the tower/Assets/Scripts/GameData/InGameData.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 public class InGameData
 {
 	public int kills;
-	public float timeWithOverdrive;
+	public double timeWithOverdrive;
 }
